leonardoai-scraper/requirements.txt
scrapy-playwright==0.0.32
pillow==10.1.0